requirements.txt
fastapi  
uvicorn  
azure-identity  
azure-keyvault-secrets  
requests 
spitch
azure-cognitiveservices-speech